% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode race and ethnicity or gender feature}
\usage{
recode(x, feature = "race")
}
\arguments{
\item{x}{Vector of race and ethnicity or gender values}

\item{feature}{Name of feature; currently accepts "race" or "gender"}
}
\value{
Consistent values of race and ethnicity or gender feature
}
\description{
Recode race and ethnicity or gender feature
}
\examples{
x <- c("M")
recode(x, feature = "gender")
}
