% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_consistency.R
\name{str_consistency}
\alias{str_consistency}
\title{Ensure string consistency for race and ethnicity or gender}
\usage{
str_consistency(x, feature = NULL)
}
\arguments{
\item{x}{Vector of race and ethnicity or gender values}

\item{feature}{Name of feature; currently accepts "race" or "gender"}
}
\value{
Consistent values of race and ethnicity or gender feature
}
\description{
Ensure string consistency for race and ethnicity or gender
}
\examples{
x <- c("M")
str_consistency(x, feature = "gender")
}
