% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_arrest}
\alias{tidy_arrest}
\title{Preprocess arrest data}
\usage{
tidy_arrest(file_report, file_officer)
}
\arguments{
\item{file_report}{File name in working directory or path to .csv file
containing primary arrest data, e.g. date, time, civilian attributes}

\item{file_officer}{File name in working directory or path to .csv file
containing corresponding officer data, e.g. appointed date}
}
\value{
A data.table containing joined arrest data
}
\description{
Preprocess arrest data
}
\examples{

}
