% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_assignment}
\alias{tidy_assignment}
\title{Preprocess assignment data}
\usage{
tidy_assignment(file)
}
\arguments{
\item{file}{A vector of file names in working directory or path
to .csv files containing assignment and attendance data}
}
\value{
A data.table containing assignment data
}
\description{
Preprocess assignment data
}
\examples{

}
