% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_contact}
\alias{tidy_contact}
\title{Preprocess contact card data}
\usage{
tidy_contact(file)
}
\arguments{
\item{file}{File name in working directory or path to .csv file containing
contact card data}
}
\value{
A data.table containing contact card data
}
\description{
Preprocess contact card data
}
\examples{

}
