% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_force}
\alias{tidy_force}
\title{Preprocess use of force data}
\usage{
tidy_force(file_report, file_action)
}
\arguments{
\item{file_report}{File name in working directory or path to .csv file
containing primary tactical response report data, e.g. date, time, officer}

\item{file_action}{File name in working directory or path to .csv file
containing corresponding action data, e.g. escort holds}
}
\value{
A data.table containing joined tactical response report data
}
\description{
Preprocess use of force data
}
\examples{

}
