% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{tidy_isr}
\alias{tidy_isr}
\title{Preprocess investigatory stop report data}
\usage{
tidy_isr(file)
}
\arguments{
\item{file}{File name in working directory or path to .csv file containing
investigatory stop report data}
}
\value{
A data.table containing investigatory stop report data
}
\description{
Preprocess investigatory stop report data
}
