% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest.R
\name{tidy_arrest}
\alias{tidy_arrest}
\title{Preprocess arrest data}
\usage{
tidy_arrest(file_report, file_officer)
}
\arguments{
\item{file_report}{A vector of file names in working directory or path
to .csv files containing primary arrest data, e.g. FOIA p701162}

\item{file_officer}{A vector of file names in working directory or path
to .csv file containing corresponding officer data, e.g. FOIA p708085}
}
\value{
A data.table containing joined arrest data
}
\description{
Preprocess arrest data
}
