% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_ticket}
\alias{tidy_ticket}
\title{Preprocess ticket data}
\usage{
tidy_ticket(file, zip = FALSE)
}
\arguments{
\item{file}{File name in working directory or path to .csv file containing
ticket data}

\item{zip}{logical (default is \code{FALSE}). Set to \code{TRUE} if the .csv
file is zipped and file will be unzipped prior to reading.}
}
\value{
A data.table containing ticket data
}
\description{
Preprocess ticket data
}
\examples{
file <- "parking_tickets.csv"
tidy_ticket(
  system.file("extdata", file, package = "cpdata", mustWork = TRUE)
)
}
